requirements.txt
asgiref==3.11.0
diff-match-patch==20241021
Django==5.2.8
django-import-export==4.3.14
django-unfold==0.72.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.3
tablib==3.9.0
